#262 JWT `ttl` is overflowing

+++ extensions/servlet/src/main/java/com/stormpath/sdk/servlet/config/Config.java
@@ -63,7 +63,7 @@ public interface Config extends Map<String, String> {
 
     CookieConfig getAccountCookieConfig();
 
-    int getAccountJwtTtl();
+    long getAccountJwtTtl();
 
     <T> T getInstance(String classPropertyName) throws ServletException;
